passing corner tests
